DEV-34: add user check to user routes

+++ routes/user.js
@@ -1,5 +1,5 @@
 require("dotenv").config();
-const { returnData, errorHandler } = require("./helperMethods.js");
+const { returnData, errorHandler, forbiddenHandler } = require("./helperMethods.js");
 const users = require("../model/User.js");
 const plans = require("../model/Plan.js");
 const { createToken, auth } = require("../util/token");
@@ -13,7 +13,7 @@ const router = express.Router();
 
 const DEBUG = process.env.DEBUG === "True";
 
-router.get("/api/user", auth, (req, res) => {
+router.get("/api/user", (req, res) => {
   const username = req.query.username || "";
   const affiliation = req.query.affiliation || "";
   const query = {
@@ -66,8 +66,12 @@ router.get("/api/backdoor/verification/:id", (req, res) => {
   });
 });
 
+// NOT_IN_USE
 router.delete("/api/user/:id", auth, async (req, res) => {
   const id = req.params.id;
+  if (req.user._id !== id) {
+    return forbiddenHandler(res);
+  }
   const user = await users.findByIdAndDelete(id);
   if (user) {
     await courses.deleteMany({ user_id: id });
